// frontend/TimezoneManager -TypeScript/src/types/store/actionTypes.tsx
export enum UserActionTypes {
    AUTH_AUTO_SIGN_IN = 'AUTH_AUTO_SIGN_IN',
    AUTH_USER = 'AUTH_USER',
    AUTH_USER_SUCCESS = 'AUTH_USER_SUCCESS',
    REGISTER_USER = 'REGISTER_USER',
    REGISTER_USER_SUCCESS = 'REGISTER_USER_SUCCESS',
    UPDATE_USER_INFO = 'UPDATE_USER_INFO',
    UPDATE_USER_INFO_SUCCESS = 'UPDATE_USER_INFO_SUCCESS',
    DELETE_USER = 'DELETE_USER',
    DELETE_USER_SUCCESS = 'DELETE_USER_SUCCESS',
    LOGOUT = 'LOGOUT',
    LOGOUT_SUCCESS = 'LOGOUT_SUCCESS',
    CHANGE_USER_ROLE = 'CHANGE_USER_ROLE',
    CHANGE_USER_ROLE_SUCCESS = 'CHANGE_USER_ROLE_SUCCESS',
    REFRESH_TOKEN_SUCCESS = 'REFRESH_TOKEN_SUCCESS'
}

export enum UIActionTypes {
    START_ACTION = 'START_ACTION',
    STOP_ACTION = 'STOP_ACTION',
    REFRESH_ACTION_START = 'REFRESH_ACTION_START',
    REFRESH_ACTION_STOP = 'REFRESH_ACTION_STOP',
    TOGGLE_POPUP_MESSAGE = 'TOGGLE_POPUP_MESSAGE',
    ADD_ERROR_ACTION = 'ADD_ERROR_ACTION',
    CLEAR_ERROR_ACTIONS = 'CLEAR_ERROR_ACTIONS'
}

export enum TimezoneActionTypes {
    ADD_NEW_TIMEZONE_ENTRY = 'ADD_NEW_TIMEZONE_ENTRY',
    ADD_NEW_TIMEZONE_ENTRY_SUCCESS = 'ADD_NEW_TIMEZONE_ENTRY_SUCCESS',
    UPDATE_TIMEZONE_ENTRY = 'UPDATE_TIMEZONE_ENTRY',
    UPDATE_TIMEZONE_ENTRY_SUCCESS = 'UPDATE_TIMEZONE_ENTRY_SUCCESS',
    DELETE_TIMEZONE_ENTRY = 'DELETE_TIMEZONE_ENTRY',
    DELETE_TIMEZONE_ENTRY_SUCCESS = 'DELETE_TIMEZONE_ENTRY_SUCCESS',
    FETCH_USER_TIMEZONE_ENTRIES = 'FETCH_USER_TIMEZONE_ENTRIES',
    FETCH_USER_TIMEZONE_ENTRIES_SUCCESS = 'FETCH_USER_TIMEZONE_ENTRIES_SUCCESS'
}

export enum SearchActionTypes {
    SEARCH_USER_TIMEZONE_ENTRIES = 'SEARCH_USER_TIMEZONE_ENTRIES',
    SEARCH_USER_TIMEZONE_ENTRIES_SUCCESS = 'SEARCH_USER_TIMEZONE_ENTRIES_SUCCESS',
    SEARCH_ALL_TIMEZONE_ENTRIES = 'SEARCH_ALL_TIMEZONE_ENTRIES',
    SEARCH_ALL_TIMEZONE_ENTRIES_SUCCESS = 'SEARCH_ALL_TIMEZONE_ENTRIES_SUCCESS',
    SEARCH_USERS = 'SEARCH_USERS',
    SEARCH_USERS_SUCCESS = 'SEARCH_USERS_SUCCESS',
    UPDATE_SEARCHED_USER_INFO_SUCCESS = 'UPDATE_SEARCHED_USER_INFO_SUCCESS',
    CLEAR_ALL_SEARCHES = 'CLEAR_ALL_SEARCHES',
    CLEAR_TIMEZONE_ENTRIES_SEARCH = 'CLEAR_TIMEZONE_ENTRIES_SEARCH',
    CLEAR_ALL_TIMEZONE_ENTRIES_SEARCH = 'CLEAR_ALL_TIMEZONE_ENTRIES_SEARCH',
    CLEAR_USERS_SEARCH = 'CLEAR_USERS_SEARCH',
    CHANGE_SELECTED_DROPDOWN = 'CHANGE_SELECTED_DROPDOWN',
    CHANGE_SELECTED_FILTER = 'CHANGE_SELECTED_FILTER'
}

export type StoreActionTypes = UserActionTypes | TimezoneActionTypes | SearchActionTypes;